Final project/project_silant/silant/requirements.txt
asgiref==3.6.0
certifi==2022.12.7         
cffi==1.15.1               
charset-normalizer==3.1.0  
cryptography==40.0.1       
defusedxml==0.7.1          
Django==4.1.7              
django-allauth==0.54.0     
django-filter==23.1        
djangorestframework==3.14.0
idna==3.4                  
Markdown==3.4.3            
oauthlib==3.2.2            
pycparser==2.21            
PyJWT==2.6.0               
python3-openid==3.2.0      
pytz==2023.3               
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
